Add THRESHOLD parameter to RatedUnqueue
When set, RatedUnqueue will wait until it is possible to pull at least
threshold number of packets in order to reduce runs and empty runs numbers.

Default threshold is 1. This means that RatedUnqueue will pull from queue
as soon as it is possible to pull a single packet. This can result in
single-packet batches and increased number of empty runs.