Less precise RatedUnqueue
Add PRECISE parameter to RatedUnqueue.

Default:
  - false when compiled with HAVE_BATCH
  - true when not compiled with HAVE_BATCH

When set to false, RatedUnqueue will wait until it is
possible to pull at least 32 packets in order to reduce
runs and empty runs numbers.

When set to true, RatedUnqueue will pull from queue as soon
as it is possible to pull single packet. This can result in
single-packet batches and increased number of empty runs.